% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{frame_picture}
\alias{frame_picture}
\title{Include graphical objects in frames for convenient printing.}
\usage{
frame_picture(gobj, title = "Title", border_col = "black",
  background_fill = "grey", ...)
}
\arguments{
\item{gobj}{Grob.}

\item{title}{Character(1).}

\item{border_col}{Character(1).}

\item{background_fill}{Character(1).}

\item{...}{Extra strip parameters passed to rectGrob.}
}
\value{
FramePic.
}
\description{
Include graphical objects in frames for convenient printing.
}
