% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{ob_flatten}
\alias{ob_flatten}
\title{A utility for flattening a composite file containing multiple molecules
into a bunch of separate ones.}
\usage{
ob_flatten(filename, cmd = "obabel", output = tempdir(),
  extra = "-m --unique")
}
\arguments{
\item{filename}{Character.}

\item{cmd}{Character(1).}

\item{output}{Directory(1).}

\item{extra}{Character(1).}
}
\description{
A utility for flattening a composite file containing multiple molecules
into a bunch of separate ones.
}
