% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{ob_render}
\alias{ob_render}
\title{A utility to render sdf files into Cairo SVG images so they can
be imported into R.}
\usage{
ob_render(filename, cmd = "obabel", output = tempdir(),
  extra = "-d -xC -xm -xj -xt -xd -xb none")
}
\arguments{
\item{filename}{Character file names of chemical structures.}

\item{cmd}{Character(1) obabel executable path.}

\item{output}{Character(1) directory to output files.}

\item{extra}{Character(1) arguments to format the SVG images.}
}
\description{
A utility to render sdf files into Cairo SVG images so they can
be imported into R.
}
