% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{print.framepic}
\alias{print.framepic}
\title{Print objects framepic to the graphics device.}
\usage{
\method{print}{framepic}(x)
}
\arguments{
\item{x}{framepic object.}
}
\value{
Nothing
}
\description{
Print objects framepic to the graphics device.
}
