% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{combine_frames}
\alias{combine_frames}
\title{Combine multiple frames together}
\usage{
combine_frames(fs, ...)
}
\arguments{
\item{fs}{List-Of(Frames).}

\item{...}{Extra frames which will get combined.}
}
\description{
Combine multiple frames together
}
